requirements.txt
selenium
paho.mqtt
requests